.NET core preview idea
Added a note in here about how AzDO could be used like we're using it now to publish .NET Core 3.0 Preview 7 builds of the dot.net web site. People ask a lot about how to get preview builds running on App Service, and i think AzDO makes this a breeze and it should be called out for customers.